Proper static representation and StaticRadicli
