+++ radicli/util.py
@@ -1,24 +1,17 @@
 from typing import Any, Callable, Iterable, Type, Union, Optional, Dict, Tuple
 from typing import List, Literal, NewType, get_args, get_origin, TypeVar
-from typing import TYPE_CHECKING
 from enum import Enum
 from dataclasses import dataclass
 from pathlib import Path
 import inspect
 import argparse
-import json
 
 # We need this Iterable type, which is the type origin of types.Iterable
 try:
     from collections.abc import Iterable as IterableType  # Python 3.9+
 except ImportError:
     from collections import Iterable as IterableType  # type: ignore
-if TYPE_CHECKING:
-    from . import Radicli
 
-STATIC_ROOT_KEY = "root"
-STATIC_CMD_KEY = "cmd"
-STATIC_SUB_KEY = "sub"
 DEFAULT_PLACEHOLDER = argparse.SUPPRESS
 BASE_TYPES = [str, int, float, Path]
 ConverterType = Callable[[str], Any]
@@ -253,57 +246,6 @@ def format_arg_help(text: Optional[str], max_width: int = 70) -> str:
     d = (text or "").strip()[:max_width]
     end = "." if "." in d or len(text or "") <= max_width else "..."
     return (d.rsplit(".", 1)[0] if "." in d else d) + end
-
-
-def print_static_help(
-    path: Optional[Union[str, Path]],
-    cmd: Optional[str] = None,
-    sub: Optional[str] = None,
-) -> bool:
-    if not path or not Path(path).exists() or not Path(path).is_file():
-        return False
-    with Path(path).open("r", encoding="utf8") as f:
-        data = json.load(f)
-    static = get_static_help(data, cmd, sub)
-    if static:
-        print(static)
-        return True
-    return False
-
-
-def get_static_help(
-    data: Dict[str, Any],
-    cmd: Optional[str] = None,
-    sub: Optional[str] = None,
-) -> Optional[str]:
-    if not cmd:
-        return data[STATIC_ROOT_KEY]
-    if cmd and cmd not in data[STATIC_CMD_KEY]:
-        return None
-    if sub and sub in data[STATIC_SUB_KEY][cmd]:
-        return data[STATIC_SUB_KEY][cmd][sub]
-    elif cmd:
-        return data[STATIC_CMD_KEY][cmd]
-    return None
-
-
-def generate_static_help(cli: "Radicli") -> Dict[str, Any]:
-    cmds: Dict[str, str] = {}
-    subs: Dict[str, Dict[str, str]] = {}
-    for cmd in cli.commands.values():
-        subcmds = cli.subcommands.get(cmd.name, {})
-        parser, subparsers = cli.get_parsers(
-            cmd.args, subcmds, name=cmd.name, description=cmd.description
-        )
-        cmds[cmd.name] = parser.format_help()
-        subs[cmd.name] = {}
-        for subp, subcmd in subparsers.values():
-            subs[cmd.name][subcmd.name] = subp.format_help()
-    return {
-        STATIC_ROOT_KEY: cli.format_info(),
-        STATIC_CMD_KEY: cmds,
-        STATIC_SUB_KEY: subs,
-    }
 
 
 def expand_error_subclasses(
